test: split the dashboard reader scope tests into their own file
The new cases pushed test_tenant_scope_http_isolation.py to 238 tests,
past the per-file subprocess timeout in scripts/run_tests.py, so the whole
file was killed at 600s in CI rather than reporting a result.

Move them to test_tenant_scope_dashboard_readers.py, which imports the
harness from the original rather than rebuilding it. The split is along
the seam that already existed in the file: the original covers the task
routes, the new file covers the operator-facing readers over the same
table.

Each file now runs alone in roughly three minutes and no longer trips the
runner's memory guard mid-file.
